app/requirements.txt
Django==5.0.7
requests==2.32.3
django-json-widget==2.0.1
django-extensions==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
Markdown==3.6
django-filter==24.2
python-dateutil==2.9.0.post0

pytest-django
pytest

black
flake8
pur
pylint
coverage
